refactor: move to main.go
